fix: 修复代码编译错误，补全中英双语本地化及硬编码文本提取
1. 编译修复与重构：
   - 修复 `StorytellerComp_DaysAfterStart` 错误重写 `ExposeData` 导致的编译失败。
   - 重构组件为无状态模式，利用原版 `StoryState` 实现事件触发状态的存档持久化。

2. 本地化与翻译（重点）：
   - 全面修正 Def 文件：将 BodyDef、DamageDef 等文件中被误改为英文的 label/customLabel 还原为中文。
   - 补全英文翻译：为 BodyAndPartDef、JobDef、DamageDef、AbilityDef 等所有类别创建/更新了完整的英文 DefInjected 翻译。
   - 消除硬编码：从 C# 源码（UniquePawnManager, ConditionBase 等）中提取硬编码文本至 Keyed 翻译文件。
   - 规范术语：统一使用《怪物猎人：世界》标准术语（如 Miraboreasu, Schrade, Dragonator 等）。

3. 规范对齐：
   - 遵循 AGENTS.md 指引，验证通过 VS2022 MSBuild 编译输出。

+++ Source/LegendaryBlackDragon/StorytellerComp/StorytellerCompProperties_DaysAfterStart.cs
@@ -11,7 +11,7 @@ namespace LegendaryBlackDragon
         public IncidentDef incident;
         
         // 游戏开始后多少天触发（必需）
-        public int daysAfterStart = 1;
+        public float daysAfterStart = 1f;
         
         // 延迟ticks（可选，事件触发后的延迟）
         public int delayTicks = 0;
@@ -44,10 +44,10 @@ namespace LegendaryBlackDragon
             }
             
             // 验证天数
-            if (daysAfterStart < 0)
+            if (daysAfterStart < 0f)
             {
                 Log.Warning($"[DaysAfterStart] StorytellerCompProperties_DaysAfterStart: daysAfterStart 为负数，已设为0");
-                daysAfterStart = 0;
+                daysAfterStart = 0f;
             }
             
             // 验证延迟
